sync: route SyncManager server pulls through SyncRepository (fixes #15466)
Rebased onto master, squashing the three iterations of this refactor and
reconciling it with the dashboard sync changes that landed in the meantime.

- Convert `SyncRepository` to an interface and move the implementations of
  `uploadLoginData`, `uploadBulkData`, `processShelfParallel`, and
  `processShelfDataOptimizedSync` into `SyncRepositoryImpl`.
- Keep `processShelfDataOptimizedSync` private in the impl to avoid
  over-exposing it on the interface.
- Keep master's `suspend fun syncDashboardKeyId(role: String?): SyncUiState`
  signature on the interface, implemented in `SyncRepositoryImpl` via
  `TransactionSyncManager`. `BaseDashboardFragment` keeps routing through
  `DashboardViewModel`, which already calls `SyncRepository`.
- Refactor `SyncManager` to depend on `SyncRepository` rather than executing
  shelf pull logic directly, and drop the dead `stringArray` state.
- Bind `SyncRepositoryImpl` to `SyncRepository` in `RepositoryModule`.
- Update `SyncManagerTest` to mock the new dependency.

`provideSyncManager` is already gone from `ServiceModule` on master, so this
branch no longer needs to remove it.
