+++ app/src/test/java/org/ole/planet/myplanet/services/sync/SyncManagerTest.kt
@@ -73,6 +73,7 @@ class SyncManagerTest {
             teamsSyncRepository,
             coursesRepository,
             eventsRepository,
+            mockk(),
             mockk()
         )
     }
